moved delete_collection into journalist_app/utils.py
