+++ securedrop/journalist_app/utils.py
@@ -189,3 +189,17 @@ def make_password():
             return password
         except PasswordError:
             continue
+
+
+def delete_collection(filesystem_id):
+    # Delete the source's collection of submissions
+    job = worker.enqueue(srm, store.path(filesystem_id))
+
+    # Delete the source's reply keypair
+    crypto_util.delete_reply_keypair(filesystem_id)
+
+    # Delete their entry in the db
+    source = get_source(filesystem_id)
+    db_session.delete(source)
+    db_session.commit()
+    return job
